tests/wpt/tests/css/css-transforms/parsing/transform-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 1: getComputedStyle().transform</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-1/#serialization-of-the-computed-value">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    width: 200px;
    height: 300px;
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("transform", "perspective(none)", "matrix(1, 0, 0, 1, 0, 0)");
test_computed_value("transform", "perspective(10px)", "matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, -0.1, 0, 0, 0, 1)");

// FIXME: This needs more tests to cover all the other transform functions.
</script>
</body>
</html>
